+++ src/Strategy/WeightedTactics/WeightedTacticInterface.php
@@ -6,9 +6,7 @@ use App\Model\GamePlayInterface;
 
 interface WeightedTacticInterface
 {
-    public function getWeight(GamePlayInterface $game, string $location, bool $isFallbackToHeroLocation): int;
+    public function getWeight(GamePlayInterface $game, string $location, bool $isFallbackToHeroLocation): float;
 
     public function isApplicableLocation(GamePlayInterface $game, string $location): bool;
-
-    //public function isWalkable(GamePlayInterface $game, string $location): bool;
 }
